2.8.2

+++ fw/application/src/mui/mui_canvas.h
@@ -27,7 +27,7 @@ uint8_t mui_canvas_get_utf8_bytes(const char *p);
 
     void mui_canvas_set_font(mui_canvas_t *p_canvas, const uint8_t *font);
 uint8_t mui_canvas_draw_utf8(mui_canvas_t *p_canvas, uint8_t x, uint8_t y, const char *str);
-uint8_t mui_canvas_draw_utf8_clip(mui_canvas_t *p_canvas, uint8_t x, uint8_t y, const char *text);
+int32_t mui_canvas_draw_utf8_clip(mui_canvas_t *p_canvas, int32_t x, int32_t y, const char *text);
 uint8_t mui_canvas_draw_glyph(mui_canvas_t *p_canvas, uint8_t x, uint8_t y, uint16_t encoding);
 uint16_t mui_canvas_get_utf8_width(mui_canvas_t *p_canvas, const char *s);
